+++ checks/service-dummy-test-from-flake/flake.nix
@@ -6,7 +6,7 @@
   outputs =
     { self, clan-core, ... }:
     let
-      # Usage see: https://docs.clan.lol
+      # Usage see: https://clan.lol/docs
       clan = clan-core.lib.clan {
         inherit self;