+++ nixosModules/clanCore/deprecated.nix
@@ -6,15 +6,15 @@ in
   imports = [
     # == clan.inventory ==
     (mkRemovedOptionModule [ "clan" "inventory" "services" ] ''
-      ###############################################################################
-      #                                                                             #
-      # Clan modules (clanModules) have been deprecated and removed in favor of     #
-      # Clan services!                                                              #
-      #                                                                             #
-      # Refer to https://docs.clan.lol/guides/migrations/migrate-inventory-services #
-      # for migration instructions.                                                 #
-      #                                                                             #
-      ###############################################################################
+      ########################################################################################
+      #                                                                                      #
+      # Clan modules (clanModules) have been deprecated and removed in favor of              #
+      # Clan services!                                                                       #
+      #                                                                                      #
+      # Refer to https://clan.lol/docs/unstable/guides/migrations/migrate-inventory-services #
+      # for migration instructions.                                                          #
+      #                                                                                      #
+      ########################################################################################
     '')
     (mkRemovedOptionModule [ "clan" "inventory" "assertions" ] ''
       Use of 'machine.<name>.config.clan.inventory.assertions' was replaced by 'clan.checks'